move exporter/trace/otlp to exporters/otlp
